requirements.txt
Django==3.0.7
django-ckeditor==6.1.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-payments==0.14.0
djangorestframework==3.11.1
django-bootstrap4
django-import-export==2.6.1
easy-thumbnails==2.8
django-cleanup==5.2.0
django-ckeditor==6.1.0
django-multiselectfield==0.1.12
django-simple-captcha==0.5.14
django-cors-headers==3.4.0
djangorestframework==3.11.1
